Started upgrading to 1.0
Simple things appear to work pretty well now, but there are probably still some issues to be discovered

+++ src/main/java/org/peimari/gleaflet/client/MapOptions.java
@@ -24,7 +24,7 @@ public class MapOptions extends AbstractJsonOptions {
 		this.attributionControl = b;
 	}-*/;
 
-	public native final void setZoom(int zoom) 
+	public native final void setZoom(double zoom) 
 	/*-{
 		this.zoom = zoom;
 	}-*/;
